refactor: improvements to all infra token related components
